7.1: routed anecdotes, step1
Add React Router to the application so that by clicking links in the Menu component the view can be changed.

At the root of the application, meaning the path /, show the list of anecdotes

The Footer component is always visible at the bottom.

The creation of a new anecdote happens in the path /create: